test_runner: move fakerTimers to mock module
