+++ lib/internal/test_runner/mock/mock.js
@@ -26,6 +26,7 @@ const {
   validateInteger,
   validateObject,
 } = require('internal/validators');
+const { FakeTimers } = require('internal/test_runner/mock/fake_timers');
 
 function kDefaultFunction() {}
 
@@ -35,6 +36,7 @@ class MockFunctionContext {
   #implementation;
   #restore;
   #times;
+  #fakeTimers;
 
   constructor(implementation, restore, times) {
     this.#calls = [];
@@ -48,6 +50,10 @@ class MockFunctionContext {
     return ArrayPrototypeSlice(this.#calls, 0);
   }
 
+  get fakeTimers() {
+    this.#fakeTimers ??= new FakeTimers();
+    return this.#fakeTimers;
+  }
   callCount() {
     return this.#calls.length;
   }
