+++ lib/internal/test_runner/test.js
@@ -30,8 +30,7 @@ const {
   },
   AbortError,
 } = require('internal/errors');
-const { MockTracker } = require('internal/test_runner/mock');
-const { FakeTimers } = require('internal/test_runner/fake_timers');
+const { MockTracker } = require('internal/test_runner/mock/mock');
 const { TestsStream } = require('internal/test_runner/tests_stream');
 const {
   createDeferredCallback,
@@ -107,11 +106,6 @@ class TestContext {
   get mock() {
     this.#test.mock ??= new MockTracker();
     return this.#test.mock;
-  }
-
-  get fakeTimers() {
-    this.#test.fakeTimers ??= new FakeTimers();
-    return this.#test.fakeTimers;
   }
 
   runOnly(value) {
