+++ lib/test.js
@@ -23,28 +23,11 @@ ObjectDefineProperty(module.exports, 'mock', {
   enumerable: true,
   get() {
     if (lazyMock === undefined) {
-      const { MockTracker } = require('internal/test_runner/mock');
+      const { MockTracker } = require('internal/test_runner/mock/mock');
 
       lazyMock = new MockTracker();
     }
 
     return lazyMock;
-  },
-});
-
-let lazyFakeTimers;
-
-ObjectDefineProperty(module.exports, 'fakeTimers', {
-  __proto__: null,
-  configurable: true,
-  enumerable: true,
-  get() {
-    if (lazyFakeTimers === undefined) {
-      const { FakeTimers } = require('internal/test_runner/fake_timers');
-
-      lazyFakeTimers = new FakeTimers();
-    }
-
-    return lazyFakeTimers;
   },
 });
